refactor: 진행도 컴포넌트 작게 분할
